JAMES-2586 Moving Managers out of the mail package
Matching logic with tests package location and cassandra mailbox module
